+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### New Features 🎉
 
 - Implement the `$slice`, `$sort` and `$position` modifiers for the `$push` update operator (used by WeKan) by @xet7. Thanks to xet7.
+- Implement the `$eq`, `$ne`, `$or`, `$ifNull`, `$anyElementTrue`, `$objectToArray` and `$map` aggregation expression operators used by WeKan by @xet7. Thanks to xet7.
 
 ### Other Changes 🤖